fix: Strip console codes from RSpec messages and better handle multiple failures (#1232)
* fix: strip console codes from failure messages

The instrumentation was including console codes in failure messages,
which was leading to leading to some slightly brittle tests. We don't
need console codes so we should strip them.

* fix: handle multiple exceptions

RSpec wraps multiple failures and errors into a single exception. This
commit handles this exception to provide clearer reporting of the
failures and errors.

* fix: add a delta to span timestamp check