chore: Add rake task to validate inter-repo dependencies
This commit creates a task that parses gemspecs for each gem in the
repo, and then validates that each opentelemetry dependency for every
gem can be satisfied by the versions that are in the repo. The intent is
to ensure that we aren't releasing gems that will break because of
dependency failures (overly pessimistic specifications, etc).

To do this, we slightly modify the existing tasks:
- The `foreach_gem` method expects a block, but still changes into the
  gem's directory before yielding the gem name to the block.
- We add a new `do_cmd` method that conditionally runs the requested
  shell command with or without a clean bundler env, as before.