fixup: use schema_url = nil instead of blank string for Resource#create
The spec says "empty" - which you could say is possibly an empty string
(which was what I did initially). But given that an
`InstrumentationLibrary` in this implementation is just a struct, an ILS
initialized without a schema_url will actually get one that is `nil`,
not `''`

Just better to be consistent.

We treat empty-string schema_urls as if they were opaque; don't do that
unless you really want it.