fix: Omit `nil` `net.peer.name` attributes
Faraday uses a base URL to derive the host name which is used by the tracer middleware as the span attribute value for `net.peer.name`.

Although it seems incorrect, Faraday does not require a base URL when it is initialized and is a commonly used when using Faraday test stubs.

When the base URL is absent, the `net.peer.name` is set to `nil` triggering and invalid attribute error report and creates a bit of noise for test cases.

This change omits the `net.peer.name` attribute when it is `nil` to suppress error reporting when the host is absent.