feat: add schema_url to instrumentation_base and resource_detectors
Any instrumentation can now set an `instrumentation_schema_url` if they
so choose. Resource detectors may also do so.

Notably, for the GCP resource detector, we explicitly pass a `nil`
schema_url. That's the default, of course, but I made it explicit since
we haven't actually audited whether or not we conform to any schema.