fix: make sidekiq instrumentation compatible with sidekiq 6.5.0
This commit:

- changes our config method to instrumentation_config. Sidekiq's new
Component module is mixed in in a lot of places, and that module
includes a config method that we very much do not want to overwrite.

- changes the MockLoader in our test suite to be compatible with Sidekiq
6.5.0. Starting in Sidekiq 6.5.0, Sidekiq accepts the Sidekiq singleton
as an argument to Launcher.new. It seems like this is an intermediate
step to making Sidekiq::Config its own class, which is planned for
Sidekiq 7.0.

- adds appraisals for sidekiq 6.4.0 and 6.5.0