feat: support OTEL_ATTRIBUTE_{COUNT,VALUE_LENGTH}_LIMIT env vars (#1292)
Fixes #1251 by additionally supporting `OTEL_ATTRIBUTE_COUNT_LIMIT` and
`OTEL_ATTRIBUTE_VALUE_LENGTH_LIMIT` as ways to configure span limits. As
per spec, the existing model-specific limits (eg: `OTEL_SPAN_ATTRIBUTE_...`
env vars) are preferred when they are present.

The easiest way to do this was to re-use the handy `config_opt` method
in `opentelemetry-common` package, on which we already depend. It made
the lines even longer, which is gross, but it works pretty nicely.

Co-authored-by: Francis Bogsanyi <francis.bogsanyi@shopify.com>