feat: support truffleruby (#1295)
* feat: support truffleruby

Does what it says on the tin. By "support" I mean "turn it on in CI and
see if everything works."

The compatibility list is actually a bit better than JRuby!

Notably, I needed to exclude the OTLP exporter to make tests pass. I'm
not certain it's a real bug; I just haven't tried to fix those test
failures yet.

* fixup: exclude more grpc things

Co-authored-by: Francis Bogsanyi <francis.bogsanyi@shopify.com>