chore(deps): Allow google-protobuf `~> 3.14`
GitHub still uses legacy versions of the protobuf gem. This makes it challenging to
upgrade to newer versions of the OTLP Exporter which only supports versions `~> 3.19`.

This change loosens the restrictions to allow GitHub to adopt newer versions of the protobuf
definitions using an older verson of the library.

This change explicitly skips 3.15 due to bugs like https://github.com/protocolbuffers/protobuf/issues/8337