BatchSpanProcessor#force_flush: purge inherited spans even on shutdown
Ref: https://github.com/open-telemetry/opentelemetry-ruby/pull/363
Ref: https://github.com/Shopify/pitchfork/pull/74

At Shopify we're using the Pitchfork web server, and one thing it
does is that it sometimes do a "double fork", similar to the classic
daemonization dance.

Recently we noticed that sometimes the "middle process" of that
daemonization dance would get "stuck". After adding some debug
we figured it's because we have:

```ruby
at_exit do
  # Triggers flushing of buffer on shutdown
  OpenTelemetry.tracer_provider.shutdown(timeout: 60)
end
```

So the child process inherits lots of spans from its parents,
and normally OpenTelemetry deal gracefully with that by droping
the inherited spans.

But here because we called `shutdown`, `@keep_running` is set to `false`
hence `reset_on_fork` is not called.

I don't understand why we shouldn't call it during shutdown though, aside
from avoiding to respawn the thread.

With this change we expect our "middle process" to drop all the inherited
spans and to have nothing to flush.