fix: Add csv as a dependency for Ruby 3.4
csv has been a default gem for a long while. In Ruby 3.4, it will stop
being a default gem and will need to be explicitly required as a
dependency.

This change adds it as a dependency for the OTLP exporter since it makes
use of the library. It uses the same version constraint as the Zipkin
exporter for consistency.