fix: Remove measure_request_duration
The OTLP metrics exporter does not use the metrics reporter present in
the traces OTLP exporter. The measure_request_duration method is used
in the traces exporter to provide a value for the metrics reporter.
Since the metrics OTLP exporter does not use this metrics reporter,
the method is not needed.

Also, with the current code, the duration was being calculated in a void
context. The value of the response variable has been and should be
the result of @http.request(request).