fix: Remove Metrics OTLP exporter `Util#measure_request_duration` and duplicate files (#1717)
* fix: Remove measure_request_duration

The OTLP metrics exporter does not use the metrics reporter present in
the traces OTLP exporter. The measure_request_duration method is used
in the traces exporter to provide a value for the metrics reporter.
Since the metrics OTLP exporter does not use this metrics reporter,
the method is not needed.

Also, with the current code, the duration was being calculated in a void
context. The value of the response variable has been and should be
the result of @http.request(request).

* fix: Remove old files outside nesting structure

There were some duplicate files at the path exporter/otlp/* that
were no longer used, now that the files are nested under
exporter/otlp/metrics/*

The test file location was updated to reflect the new location.

* fix: Remove measure_request_duration from Util

* style: Rubocop

* style: Rubocop