test: Add workaround for windows test failure
We have to instantiate the processor outside of the stub so that the
stub can be applied to the processor, but the method we're stubbing
runs on initialization, so we sometimes run out of expectations.

If we explicitly set 'OTEL_RUBY_BLRP_START_THREAD_ON_BOOT' to 'false'
for the 'removes the older log records from the batch if full' test, the
stubbed `work` method will not run outside of the stub block.

Co-authored-by: Tanna McClure <tannalynn@users.noreply.github.com>