fix: handle float value in NumberDataPoint (#1767)
* Fix: handle float value in NumberDataPoint

It's currently possible to have a `Counter` with a `Float` value, but
this fails to export because the `NumberDataPoint` is always assumed to
be an `Integer` for the purposes of ProtoBuf encoding. The `.proto` for
a `NumberDataPoint` allows for a float value in the `as_double` key, so
there should be no reason that we should not be able to support `Float`
`Counter`s.

This commit adds support for `Float` values in `NumberDataPoint`, which
in turn adds support for `Float` `Counter`s. I also updated the
documentation in the structs for both `NumberDataPoint.value` and
`HistogramDataPoint.sum` to match the expected type based on the
`.proto`.

Adds a test to confirm that the metrics_exporter handles both `Integer`
and `Float` values for a `NumberDataPoint`.

* Update metrics_sdk/lib/opentelemetry/sdk/metrics/aggregation/histogram_data_point.rb

Co-authored-by: Kayla Reopelle <87386821+kaylareopelle@users.noreply.github.com>

---------

Co-authored-by: Kayla Reopelle <87386821+kaylareopelle@users.noreply.github.com>