docs: fix Resource merge documentation (#1878)
`OpenTelemetry::SDK::Resources::Resource#merge` uses `Hash#merge` under the hood, which overwrites the "current" hash's value with the attributes of the `other` resource, so the docs at present are incorrect. It's possible this is per the spec and the implementation is wrong, but if so, it's likely easier to fix that once they're both aligned.

Co-authored-by: Kayla Reopelle <87386821+kaylareopelle@users.noreply.github.com>