ci: Update release workflows to use otelbot
Using a bot token will allow us to have CI workflows run automatically
when a release PR opens. The current setup requires us to add a commit
to get the PR to run.
This aligns the core repo with contrib.

Also, update the release author to be the new bot.